refactor: Remove `Library` class
BREAKING CHANGE:

This removes the `Library` class, which - among all known dependencies -
is not used. The dependencies checked were:

   - chai
   - deep-eql
   - assertive-chai
   - boolify-string
   - festivals-model
   - params
   - run-yo
   - react-form
   - sorrow
   - nwjs-open-link-in-browser
   - tape-chai
   - manticore
   - geospatial
   - react-form-inputs
   - seeif
   - fuzzur
   - lackey-form-data
   - compare-errors
   - node-consumer-pact-validation
   - gaia-schema
   - compare-erros

As none of these dependencies use `Library` in any way, we can easily
remove it without impacting the ecosystem, and on the plus side - we
can focus on making type-detect's core utility as good as it can be.