test: Fix tests running in Node 6.0
Symbol.toStringTag is an available read only property in Node 6. The tests overwrite the property
and as such were failing in Node 6.0. This fixes the tests so that they don't assign on Symbol
if the property already exists.

It also alters the tests to use Symbol.toStringTag proper, rather than the sham string - so that
when toStringTag actually exists, it is actually used rather than the sham; allowing tests to pass
in environments which support toStringTag for real.